pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bank-client-subscription-classifier"
version = "0.1.0"
description = "Kaggle Playground Series S5E8 - Bank Client Subscription Prediction"
readme = "README.md"
requires-python = ">=3.13.5"
license = {text = "MIT"}
authors = [
    {name = "Álvaro Granados", email = "alvarog2491@gmail.com"}
]
keywords = ["machine-learning", "mlflow", "kaggle", "classification"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "mlflow>=2.8.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "plotly>=5.15.0",
    "lightgbm>=4.0.0",
    "xgboost>=2.0.0",
    "catboost>=1.2.0",
    "shap>=0.43.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.4.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings>=0.23.0",
    "mkdocstrings-python>=1.7.0",
]

[project.urls]
Homepage = "https://alvaro-ai-ml-ds-lab.com/bank-client-subscription-classifier"
Repository = "https://github.com/alvarog2491/bank-client-subscription-classifier"
Issues = "https://github.com/alvarog2491/bank-client-subscription-classifier/issues"

[tool.black]
line-length = 79
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | mlruns
  | mlartifacts
)/
'''

[tool.flake8]
max-line-length = 79
extend-ignore = [
    "E203",  # whitespace before ':'
    "W503",  # line break before binary operator
]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "mlruns",
    "mlartifacts",
    "*.egg-info",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"